+++ engine/apps/api/views/webhooks.py
@@ -1,14 +1,23 @@
 from django.apps import apps
 from django.core.exceptions import ObjectDoesNotExist, PermissionDenied
+from django_filters import rest_framework as filters
+from rest_framework.decorators import action
 from rest_framework.exceptions import NotFound
+from rest_framework.filters import SearchFilter
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 
 from apps.api.permissions import RBACPermission
 from apps.api.serializers.webhook import WebhookSerializer
 from apps.auth_token.auth import PluginAuthentication
 from apps.webhooks.models import Webhook
+from common.api_helpers.filters import ByTeamModelFieldFilterMixin, ModelFieldFilterMixin, TeamModelMultipleChoiceFilter
 from common.api_helpers.mixins import PublicPrimaryKeyMixin, TeamFilteringMixin
+
+
+class WebhooksFilter(ByTeamModelFieldFilterMixin, ModelFieldFilterMixin, filters.FilterSet):
+    team = TeamModelMultipleChoiceFilter()
 
 
 class WebhooksView(TeamFilteringMixin, PublicPrimaryKeyMixin, ModelViewSet):
@@ -17,6 +26,7 @@ class WebhooksView(TeamFilteringMixin, PublicPrimaryKeyMixin, ModelViewSet):
 
     rbac_permissions = {
         "metadata": [RBACPermission.Permissions.OUTGOING_WEBHOOKS_READ],
+        "filters": [RBACPermission.Permissions.OUTGOING_WEBHOOKS_READ],
         "list": [RBACPermission.Permissions.OUTGOING_WEBHOOKS_READ],
         "retrieve": [RBACPermission.Permissions.OUTGOING_WEBHOOKS_READ],
         "create": [RBACPermission.Permissions.OUTGOING_WEBHOOKS_WRITE],
@@ -27,6 +37,10 @@ class WebhooksView(TeamFilteringMixin, PublicPrimaryKeyMixin, ModelViewSet):
 
     model = Webhook
     serializer_class = WebhookSerializer
+
+    filter_backends = [SearchFilter, filters.DjangoFilterBackend]
+    search_fields = ["public_primary_key", "name"]
+    filterset_class = WebhooksFilter
 
     def get_queryset(self, ignore_filtering_by_available_teams=False):
         queryset = Webhook.objects.filter(
@@ -48,9 +62,8 @@ class WebhooksView(TeamFilteringMixin, PublicPrimaryKeyMixin, ModelViewSet):
         # use this method to get the object from the whole organization instead of the current team
         pk = self.kwargs["pk"]
         organization = self.request.auth.organization
-
         try:
-            obj = organization.webhooks.get(public_primary_key=pk)
+            obj = organization.webhooks.filter(*self.available_teams_lookup_args).get(public_primary_key=pk)
         except ObjectDoesNotExist:
             raise NotFound
 
@@ -83,3 +96,22 @@ class WebhooksView(TeamFilteringMixin, PublicPrimaryKeyMixin, ModelViewSet):
         )[0]
         if self.request.auth.organization.pk not in enabled_webhooks_2_orgs.json_value["org_ids"]:
             raise PermissionDenied("Webhooks 2 not enabled for organization. Permission denied.")
+
+    @action(methods=["get"], detail=False)
+    def filters(self, request):
+        filter_name = request.query_params.get("search", None)
+        api_root = "/api/internal/v1/"
+
+        filter_options = [
+            {
+                "name": "team",
+                "type": "team_select",
+                "href": api_root + "teams/",
+                "global": True,
+            },
+        ]
+
+        if filter_name is not None:
+            filter_options = list(filter(lambda f: filter_name in f["name"], filter_options))
+
+        return Response(filter_options)